$(el).toggleClass("foo", true) should _always_ add the class
The 'switch' parameter does not indicate whether to toggle the class
(which would be pretty useless): rather, it indicates whether the class
should be added or removed. The jQuery documentation is a little bit
ambiguous on this point, but a close reading of the source and some
behavioural tests show this second interpretation to be correct.