prevent tap event to be triggered while scrolling through list of tappable elements
The tap event was always triggered after scrolling. So the element
which was under the finger after scrolling stops was triggered with a
'tap'. Now the positions of 'touchstart' and 'touched' are compared.
Only if they are within an given threshold of 10 px, 'tap' is executed.
Thanks to Ryan Fioravanti and (errr) google :
http://code.google.com/intl/de-DE/mobile/articles/fast_buttons.html