Commit to reduce the size of Zepto. We're now using the official MDN String.prototype.trim polyfill, as opposed to our own home- baked one.
My change was in part based on an issue in the uglifyjs compiler.
Closure uses inferred types to make "safe" changes like this:

if (x === undefined) {
    x = foo();
}

to:

x = x || foo();

UglifyJS doesn't make these changes. We might want to consider
moving to Closure compiler, if our aim is to produce the smallest js
possible.