Add $.Deferred implementation and test suite
Current implementation depends on $.Callbacks as jQuery's does. For
$.Callbacks to make sense, it's also necessary to get $.ajax to leverage
it, so that we can avoid code duplication.

I've taken jQuery's tests and adapted them to Evidence, and made an
attempt to give them a little more atomicity (for the most part).

Not implemented yet is `$.Deferred.when()`, which shall be forthcoming.

I'm hoping this can be reviewed so I can get some critiques on ways to
improve performance and shrink this code more.