Fixes #647: Update android regex to match Firefox for Android.
Firefox for Android doesn't include Android version information
in the UA string. This change allows the regex to detect Firefox
for Android as an Android browser with the caveat that os.version
will be undefined. But this can still be useful when trying to
distinguish between ios and android, for example.