Fix checking whether `event.preventDefault()` was called in IE10
Post-trigger, we usually want to check whether any of the event handlers
called `preventDefault()`. We did that by reading the standard
`defaultPrevented` property, but in IE10 its value gets reset
post-dispatch, so it's unreliable.

This makes it so that all events created by `$.Event()` or passed to
`trigger()` get extended with the `isDefaultPrevented()` method that
reflects this state reliably across browsers.