Don't send Ajax `data` query params in GET request bodies
For GET requests, the value of `data` gets appended to the URL as query
params. However, the serialized params string was also erraneously
passed as request body to `xhr.send()`.

This resets the `data` option to null after appending it to the URL. As
a consequence, the value of `data` is no longer available in
`beforeSend` callback for GET requests since the URL is already
constructed and it will be too late to change it. To change query params
in this callback, you must edit the `url` directly.