Fix JSONP callbacks for errored responses on Android
Android 2.3 doesn't call the `onerror` handlers for HTTP 500 responses
like it does for 404s. However, it does trigger the `onload` handler.

This rewrites the JSONP implementation to use a combination of `onload`
and `onerror` handlers to invoke the standard Ajax callbacks. If the
JSON data was never received, this always treats it as an error even if
it executed through `onload`.

References #871