Include modules matrix, make docs work on IE 10.
Uses a custom build of Zepto 1.1 that includes the detect, fx and
touch modules.

As Zepto now supports the two latest versions of IE
removed the `document.write` jQuery fallback in favor of nothing—
if people need old IE support they'll need to write a custom solution
(and probably use jQuery 1.x).

Added the modules matrix to be more complete. The links are wrong
and should point to respective sections in the documentation.
We might want to add skeleton sections for modules we haven't
documented yet.

This patch includes preliminary compatibility with Windows Phone 8.
Still doesn't work 100% right there yet, but it's mostly cosmetic
issues now.