Fix detecting whether JSONP was requested via query param
If `=?` characters appeared in the URL, Zepto would automatically go
into JSONP mode. However, this overly eager detection bites back when
the actual path of the URL ends with `=` character, e.g. when the URL
contains base64-encoded data.

This changes the detection so that `=?` only trigger JSONP mode if they
appear after another `?` character, to ensure they are indeed contained
in the query string portion of the URL.

Fixes #903