Threshold on touchmove before cancelling longTap
Ostensibly this is to fix an issue on Blackberry10. On that platform,
the `touchmove` event was always firing causing `cancelLongTap()` to
fire. I knew that HammerJS long press events worked in bb10 and they
were setting a threshold. After adding a similar threshold to the
`touchmove` event, Zepto `longTap` events began to work as well.

I have tested this against numerous devices and browsers, including
those on iOS, Android, Chrome, Firefox and Safari. All touch tests
passed in every environment.