Remove deprecated `load/unload` event shorthands
They were deprecated in jQuery 1.8 and conflict with Ajax `fn.load()`
method. Depending on the order of Zepto modules, `load()` was an
entirely different method, so to resolve the conflict we remove the
deprecated shorthands for event handling.

Fixes loading "event" module after "ajax" and not having Ajax `load()`
method overriden.