Enable AMD-compatible module output
When a module loader is present at runtime, this wraps Zepto as an
anonymous module. It still exports `Zepto` and `$` to the global
namespace when imported, though, to keep compatibility with plugins
that expect it there.

With no module loader, zepto.js executes normally.

Opt out of this AMD shim by setting `NOAMD=1` environment variable.