Fix ready() on IE (again)
We shouldn't use "DOMContentLoaded" event on IE10 as well as
"interactive" state. This commit is continuation of effort in PR #1235
and hopefully now it really fixed for any configuration.

With my server configuration (async script loading) ready() works on
IE9, but doesn't work on IE10 from time to time. I found that ready()
doesn't call the callback when invoked in "interactive" state. That
means the moment when "DOMContentLoaded" very likely was already fired.
In this case the callback won't invoked at all.

Current implementation can be explained in very simple words. We invoke
the callback when:

* (old IE) in "complete" state or when "load" event occurs;
* (others) in "interactive" state or when "DOMContentLoaded" event
  occurs.

The change tested with Microsoft's VM images from [1] on IE9 and IE10.
Seems to be good.

[1]: https://developer.microsoft.com/en-us/microsoft-edge/tools/vms/

Build size
----------

Calculated for a build with the default modules set:

```
zepto.js      58755 -> 58921 (+166 bytes)
zepto.min.js  26453 -> 26486 ( +33 bytes)
zepto.min.gz   9806 ->  9824 ( +18 bytes)
```

Note: Size of zgipped version is little flaky as depends on the first
line (comment) with a commit id, so don't worry if it differs on +/- 5
bytes from my numbers.